property/requirements.txt
requests
flask
flask-lambda-python36
Werkzeug==2.0.3
Flask-CORS
psycopg2-binary